requirements.txt
grpcio
grpcio-tools
protobuf
faster-whisper
llama-cpp-python
ffmpeg-python
python-dotenv
grpcio-health-checking
groq
soundfile